+++ features/support/hooks.ts
@@ -65,6 +65,8 @@ Before({ tags: "not @no-default-plugin" }, async function () {
             plugins: [createEsbuildPlugin(config)]
           })
         );
+
+        return config;
       }
     `
   );